requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
mysql-connector-python
pymongo
passlib[bcrypt]
python-jose[cryptography]
ollama
faster-whisper
ffmpeg-python
reportlab
python-docx
pdfminer.six
aiofiles
pydantic
python-dotenv
